+++ ordinarylight/compute/__init__.py
@@ -4,8 +4,10 @@ from ._core import (
     ComputeBuffer, ComputeStep, WebGpuBufferView, WebGpuComputeSequence,
     WebGpuComputeSession,
 )
+from .vulkan import VulkanBufferView, VulkanComputeSequence
 
 __all__ = [
     "ComputeBuffer", "ComputeStep", "WebGpuBufferView", "WebGpuComputeSequence",
     "WebGpuComputeSession",
+    "VulkanBufferView", "VulkanComputeSequence",
 ]